+++ lib/torch/nn/adaptive_max_pool1d.rb
@@ -1,0 +1,9 @@
+module Torch
+  module NN
+    class AdaptiveMaxPool1d < AdaptiveMaxPoolNd
+      def forward(input)
+        F.adaptive_max_pool1d(input, @output_size) #, @return_indices)
+      end
+    end
+  end
+end